+++ front-end/README.md
@@ -1,4 +1,4 @@
-# Front
+# FrontEnd
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 1.3.2.
