+++ front-end/e2e/app.e2e-spec.ts
@@ -1,6 +1,6 @@
 import { AppPage } from './app.po';
 
-describe('front App', () => {
+describe('front-end App', () => {
   let page: AppPage;
 
   beforeEach(() => {
