+++ front-end/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "front",
+  "name": "front-end",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
