+++ front-end/src/app/app.component.html
@@ -1,3 +1,3 @@
 <!--The content below is only a placeholder and can be replaced.-->
-<app-table>Loading...</app-table>
-<!-- <app-tabela></app-tabela> -->
+<app-table></app-table>
+
