+++ front-end/src/app/app.module.ts
@@ -1,16 +1,13 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import{ HttpModule } from '@angular/http'
+import { HttpModule } from '@angular/http';
 import { AppComponent } from './app.component';
 import { TableComponent } from './table/table.component';
-
-import { DataService } from './services/data.service';
-import { TabelaComponent } from './tabela/tabela.component'
+import { DataService } from './services/data.service'
 @NgModule({
   declarations: [
     AppComponent,
-    TableComponent,
-    TabelaComponent
+    TableComponent
   ],
   imports: [
     BrowserModule,
