+++ front-end/src/app/table/table.component.html
@@ -1,0 +1,26 @@
+<div class="container">
+
+    <form class="form-inline" (ngSubmit)="buscar(nameInput)">
+        <label class="sr-only" for="nameInput">Name</label>
+        <input type="text" class="form-control mb-2 mr-sm-2 mb-sm-0" #nameInput id="nameInput" placeholder="Jane Doe">      
+        <button type="submit" class="btn btn-primary">Submit</button>
+      </form>
+      
+
+  <table class="table table-striped">
+    <thead>
+      <tr>
+        <th>UUID</th>
+        <th>Name</th>
+        <th>Username</th>
+      </tr>
+    </thead>
+    <tbody>
+      <tr *ngFor="let register of registers">
+        <td>{{register.id}}</td>
+        <td>{{register.name}}</td>
+        <td>{{register.username}}</td>
+      </tr>
+    </tbody>
+  </table>
+</div>
