+++ front-end/src/favicon.ico
