+++ front-end/src/tsconfig.app.json
