+++ front-end/src/typings.d.ts
