+++ front-end/tsconfig.json
