--- front/src/app/tabela/tabela.component.html
@@ -1,17 +1,0 @@
-<table class="table">
-  <thead>
-    <tr>
-      <th>ID</th>
-      <th>First Name</th>
-      <th>Username</th>
-    </tr>
-  </thead>
-  <tbody>
-    <tr *ngFor="let register of reg">
-      <th> {{register.id}} </th>
-      <td>{{register.name}}</td>
-      <td>{{register.username}}</td>
-    </tr>
-  </tbody>
-</table>
-</div>
