--- front/src/app/table/table.component.css
@@ -1,3 +1,0 @@
-.r1{
-    padding: 20px;
-}
