--- front/src/app/table/table.component.html
@@ -1,28 +1,0 @@
-<div class="container">
-  <div class="r1">    
-    <form class="form-inline justify-content-md-center" (keydown.enter)="$event.preventDefault()">
-      <label class="sr-only" for="nameInput">Name</label>
-      <input type="text" name="name" class="form-control mb-2 mr-sm-2 mb-sm-0" #nameInput placeholder="Jane Doe" (keyup)="onSearch($event.target.value)">   
-    </form>
-  </div>
-
-
-
-
-  <table class="table">
-    <thead>
-      <tr>
-        <th>ID</th>
-        <th>First Name</th>
-        <th>Username</th>
-      </tr>
-    </thead>
-    <tbody>
-      <tr *ngFor="let register of reg">
-        <th> {{register.id}} </th>
-        <td>{{register.name}}</td>
-        <td>{{register.username}}</td>
-      </tr>
-    </tbody>
-  </table>
-</div>
